requirements.txt
Django==3.2.7
psycopg2-binary==2.9.1
gunicorn==20.1.0
djangorestframework==3.12.4
